requirements.txt
audio_recorder_streamlit
langdetect
streamlit
torch
torchvision
transformers
numpy
openai
python-dotenv
